Refactor master slave integrate test cases 3th.
